Separate data to other files
