+++ src/History.js
@@ -6,40 +6,8 @@ import { jsx, css } from '@emotion/core'
 import { breakpoints, mq } from './variables'
 import Section from './components/Section';
 import Heading from './components/Hgroup';
-import _History from './components/History'
-
-const items = [
-  {
-    age: 2011,
-    name: 'dotinstall',
-    url: 'http://dotinstall.com/users/shgtkshruch',
-    text: ['HTML・CSS・JavaScriptなどWEBテクノロジーの基礎を学ぶ。', '動画完了数は1500本以上。']
-  },
-  {
-    age: 2013,
-    name: 'Schoo',
-    url: 'https://schoo.jp/',
-    text: ['課題制作で実際にWEBサイトを制作。', '授業内のコンペティションで最優秀賞を受賞。']
-  },
-  {
-    age: 2015,
-    name: 'Lancers',
-    url: 'https://www.lancers.jp/profile/sh19e',
-    text: ['コーポレートサイトやランディングページなどの制作を受注。', '全ての案件の評価で☆5つを獲得。']
-  },
-  {
-    age: 2016,
-    name: 'Chogenbashugi',
-    url: 'http://chogenba.com/',
-    text: ['WEB制作会社の案件を元に実践的なコーディング技術を学ぶ。']
-  },
-  {
-    age: 2017,
-    name: 'Cykinso',
-    url: 'https://cykinso.co.jp/',
-    text: ['腸内フローラ解析のスタートアップに入社。', 'Ruby on Railsで自社サービスの開発に携わる']
-  }
-]
+import _History from './components/history/';
+import items from './components/history/data';
 
 const List = styled.ul`
   display: inline-block;
