+++ src/components/Link.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import styled from '@emotion/styled'
+
 import { mq } from '../variables'
 
 const Link = styled.a`
