+++ src/components/skill/index.js
@@ -1,10 +1,12 @@
 import React from 'react'
 import styled from '@emotion/styled'
+
 import Tippy from '@tippy.js/react'
 import 'tippy.js/dist/tippy.css';
 import 'tippy.js/animations/shift-toward-subtle.css';
 import 'tippy.js/themes/material.css';
-import { mq } from '../variables'
+
+import { mq } from '../../variables'
 
 const nthChildAnimation = Array.from('_'.repeat(25)).reduce((res, _, i) => {
   const delay = 0.05 * (i + 1);
